Modernize Xcode project. Quiet various warnings.

+++ Source/OPNetModules/Posix/TCP_IP/tcp_module.h
@@ -177,7 +177,7 @@
 	};
 
 	struct available_game_data {
-		long host;
+		int host;
 		word port;
 		word flags;
 		long ticks_at_last_response;
